+++ lib/types/Menu.d.ts
@@ -1,6 +1,6 @@
 import MenuItem from './MenuItem';
-import MenuOption from './types/MenuOption';
 import Panel, { PanelPosition } from './Panel';
+import MenuOption from './types/MenuOption';
 export default class Menu<Payload> extends Panel {
     ul: HTMLElement;
     level: number;